chore: cleans up policy MatchConditions validation  code.
Moves the code from policy_validation_matchconditions.go file into
policy_validation.go file, removing the unnecessary code.
